fix: 修复适配上游是virtfusion时，未记录额外的IPV6地址的问题

+++ scripts/build_nat_network.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # from
 # https://github.com/oneclickvirt/pve
-# 2025.05.17
+# 2025.06.09
 
 ########## 预设部分输出和部分中间变量
 
@@ -364,7 +364,8 @@ download_with_retry() {
 
 # 配置ndpresponder守护进程
 install_ndpresponder() {
-    if [ -n "$ipv6_address" ] && [ -n "$ipv6_prefixlen" ] && [ -n "$ipv6_gateway" ]; then
+    appended_file="/usr/local/bin/pve_appended_content.txt"
+    if [ -n "$ipv6_address" ] && [ -n "$ipv6_prefixlen" ] && [ -n "$ipv6_gateway" ] && if [ ! -s "$appended_file" ]; then
         if [ -f /usr/local/bin/pve_maximum_subset ] && [ "$(cat /usr/local/bin/pve_maximum_subset)" = false ]; then
             _blue "No install ndpresponder"
         elif [ "$system_arch" = "x86" ] || [ "$system_arch" = "x86_64" ]; then
@@ -671,14 +672,18 @@ iface vmbr2 inet6 static
     bridge_stp off
     bridge_fd 0
 EOF
-    if [ -f "/usr/local/bin/ndpresponder" ]; then
+    appended_file="/usr/local/bin/pve_appended_content.txt"
+    if [ -s "$appended_file" ]; then
+        sed -E 's/(# control-alias) [^[:space:]]+/\1 vmbr2/g; s/(iface) [^[:space:]]+/\1 vmbr2/g' "$appended_file" | sudo tee -a /etc/network/interfaces > /dev/null
+    else if [ -f "/usr/local/bin/ndpresponder" ]; then
         new_exec_start="ExecStart=/usr/local/bin/ndpresponder -i vmbr0 -n ${ipv6_address_without_last_segment}0/${ipv6_prefixlen}"
         file_path="/etc/systemd/system/ndpresponder.service"
         line_number=6
-        sed -i "${line_number}s|.*|${new_exec_start}|" "$file_path"
+        sudo sed -i "${line_number}s|.*|${new_exec_start}|" "$file_path"
     fi
     configure_ipv6_forwarding
 }
+
 
 # 配置IPV6转发设置
 configure_ipv6_forwarding() {
